coder/pyproject.toml
[project]
name = "coderllm-agent"
version = "0.1.0"
description = "Agent runtime for CoderLLM (LM-Studio-backed tool-using senior-dev agent)."
requires-python = ">=3.10"
dependencies = [
    "openai>=1.40",
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "PyGithub>=2.4",
    "gitpython>=3.1",
    "watchdog>=4.0",
    "pathspec>=0.12",
    "tree-sitter>=0.23",
    "tree-sitter-languages>=1.10",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8 ; platform_system != 'Darwin'",
    "faiss-cpu>=1.8 ; platform_system == 'Darwin'",
]

[project.scripts]
coder-agent = "coder.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["coder*"]
